Ensure test text files use LF line endings.
I'm setting the repository up to convert line endings to the native platform
when it's cloned. This causes problems for a couple of the tests that check
for files sizes, so I'm also configuring the relevant test file so that it
always checks out with LF line endings. Otherwise the tests fail on Windows.